fixed handling of double values, added overloads to JSON
